server/requirements.txt
dnspython==1.16.0
eventlet==0.25.1
greenlet==0.4.15
monotonic==1.5
python-engineio==3.10.0
python-socketio==4.3.1
six==1.13.0
